% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sample_list.R
\name{add_plate_position}
\alias{add_plate_position}
\title{add_plate_position}
\usage{
add_plate_position(
  S,
  x = as.character(1:9),
  y = c("A", "B", "C", "D", "E", "F"),
  plate = 1:10000,
  one.vial = c("sol"),
  plate.separate = NULL,
  sample.position.fixed = FALSE
)
}
\arguments{
\item{dat}{A data set object}

\item{id}{Scope (e.g., country codes or individual IDs)}

\item{time}{Time (e.g., time periods are given by years, months, ...)}
}
\value{
A data frame object that contains a summary of a sample that
    can later be converted to a TeX output using \code{overview_print}
}
\description{
Provides an overview table for the time and scope conditions of
    a data set
}
