% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_various_methods.R
\name{affy_normalize_loess}
\alias{affy_normalize_loess}
\title{affy_normalize_loess}
\usage{
affy_normalize_loess(
  mat,
  subset = sample(1:(dim(mat)[1]), min(c(5000, nrow(mat)))),
  epsilon = 10^-2,
  maxit = 1,
  log.it = TRUE,
  verbose = TRUE,
  span = 2/3,
  family.loess = "symmetric"
)
}
\arguments{
\item{dat}{A data set object}

\item{id}{Scope (e.g., country codes or individual IDs)}

\item{time}{Time (e.g., time periods are given by years, months, ...)}
}
\value{
A data frame object that contains a summary of a sample that
    can later be converted to a TeX output using \code{overview_print}
}
\description{
Provides an overview table for the time and scope conditions of
    a data set
}
\examples{
data(toydata)
output_table <- overview_tab(dat = toydata, id = ccode, time = year)

}
