% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_various_methods.R
\name{affy_normalize_qspline}
\alias{affy_normalize_qspline}
\title{affy_normalize_qspline}
\usage{
affy_normalize_qspline(
  x,
  target = NULL,
  samples = NULL,
  fit.iters = 5,
  min.offset = 5,
  spline.method = "natural",
  smooth = TRUE,
  spar = 0,
  p.min = 0,
  p.max = 1,
  incl.ends = TRUE,
  converge = FALSE,
  verbose = TRUE,
  na.rm = FALSE
)
}
\arguments{
\item{dat}{A data set object}

\item{id}{Scope (e.g., country codes or individual IDs)}

\item{time}{Time (e.g., time periods are given by years, months, ...)}
}
\value{
A data frame object that contains a summary of a sample that
    can later be converted to a TeX output using
}
\description{
Provides an overview table for the time and scope conditions of
    a data set
}
\examples{
data(toydata)
output_table <- overview_tab(dat = toydata, id = ccode, time = year)

}
