% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sample_list.R
\name{block_random}
\alias{block_random}
\title{block_random}
\usage{
block_random(S, group.var = NULL, nr.subj.per.block = 0, adjust.thres = 0.5)
}
\arguments{
\item{dat}{A data set object}

\item{id}{Scope (e.g., country codes or individual IDs)}

\item{time}{Time (e.g., time periods are given by years, months, ...)}
}
\value{
A data frame object that contains a summary of a sample that
    can later be converted to a TeX output using \code{overview_print}
}
\description{
Provides an overview table for the time and scope conditions of
    a data set
}
