% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_calib_curve.R
\name{compute_calib_curve}
\alias{compute_calib_curve}
\title{compute_calib_curve}
\usage{
compute_calib_curve(
  raw.file = NULL,
  concentration.file = NULL,
  dilution.nr = NULL,
  k = NULL,
  choose.normalizer = FALSE
)
}
\arguments{
\item{dat}{A data set object}

\item{id}{Scope (e.g., country codes or individual IDs)}

\item{time}{Time (e.g., time periods are given by years, months, ...)}
}
\value{
A data frame object
}
\description{
Provides an overview table for the time and scope conditions of
    a data set
}
