% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_normalization.R
\name{compute_normalization}
\alias{compute_normalization}
\title{compute_normalization}
\usage{
compute_normalization(
  datalist = NULL,
  norm.method = NULL,
  path.result = NULL,
  prefix = NULL,
  batch.wise = NULL,
  po.sample.to.use = NULL
)
}
\arguments{
\item{datalist}{A list of dataframe containing peaktable, sample info
and feature info}

\item{norm.method}{Normalization method to use, choose from c("bestis", "low_cv",
"pqn", "nomis", "qspline", "loess", "sum", "median", "combat", "limma")}

\item{path.result}{Path for export output}

\item{prefix}{Prefix for names of output files}

\item{batch.wise}{If normalize within batch only}

\item{po.sample.to.use}{Pool sample type to use for evaluating normalization
performance}
}
\value{
Normalized data in xlsx files
}
\description{
Compute normalization with various methods and export to xlsx
}
