% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_skyline_data.R
\name{convert_skyline_data}
\alias{convert_skyline_data}
\title{convert_skyline_data}
\usage{
convert_skyline_data(file = NULL, mode = NULL, path.result = NULL)
}
\arguments{
\item{file}{File of skyline exported peaktable}

\item{mode}{'pos' or 'neg' indicating positive or negative mdoe}

\item{path.result}{Path to output folder}
}
\value{
IS and Peaks table
}
\description{
Convert peaktable exported from skyline to needed format
}
