% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.qc.R
\name{create.qc}
\alias{create.qc}
\title{create.qc}
\usage{
create.qc(sampled_data = NULL, figure_height = NULL)
}
\arguments{
\item{sampled_data}{= NULL,}

\item{figure_height}{= NULL}
}
\value{
A list of summaries for all types of samples
}
\description{
Calculate summary for subsets of dataset
}
