% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_peak_table.R
\name{extract_peak_table}
\alias{extract_peak_table}
\title{extract_peak_table}
\usage{
extract_peak_table(
  x,
  rt.col = "rt",
  mz.col = "mz",
  id.col = "Identity",
  colsample.start = 8
)
}
\arguments{
\item{dat}{A data set object}

\item{id}{Scope (e.g., country codes or individual IDs)}

\item{time}{Time (e.g., time periods are given by years, months, ...)}
}
\value{
A data frame object that contains a summary of a sample that
    can later be converted to a TeX output using \code{overview_print}
}
\description{
Provides an overview table for the time and scope conditions of
    a data set
}
