% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_missing.R
\name{filter_by_missing}
\alias{filter_by_missing}
\title{filter_by_missing}
\usage{
filter_by_missing(x, method, threshold)
}
\arguments{
\item{x}{Input data frame with missing values}

\item{method}{Filter based on sample or feature, c("sample", "feature")}

\item{threshold}{Threshold for remove missings, percentage, e.g. 80}
}
\value{
A list containing a filtered dataframe and method used
}
\description{
Provides an overview table for the time and scope conditions of
    a data set
}
