% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_sample_feature.R
\name{filter_sample_feature}
\alias{filter_sample_feature}
\title{filter_sample_feature}
\usage{
filter_sample_feature(
  datalist = NULL,
  missing.sample.thres = NULL,
  missing.feature.thres = NULL,
  filter.sample.remove = NULL,
  filter.feature.remove = NULL,
  filter.is.remove = NULL
)
}
\arguments{
\item{datalist}{A datalist object}

\item{missing.sample.thres}{Threshold for keeping samples. e.g. 30 only keep
samples with < 30% missings}

\item{missing.feature.thres}{Threshold for keeping features. e.g. 30 only keep
features with < 30% missings}

\item{filter.sample.remove}{Sample.id to remove}

\item{filter.feature.remove}{Feature.id to remove}

\item{filter.is.remove}{Feature.id to remove}
}
\value{
A filtered datalist
}
\description{
Filter samples or features from peaktable
}
