% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sample_list.R
\name{generate_sample_list}
\alias{generate_sample_list}
\title{generate_sample_list}
\usage{
generate_sample_list(
  template,
  sample.initial = c("sol", "BL", "NIST", "CP", "PO"),
  sample.cycle = c("sol", "BL", "CP", "PO"),
  sample.final = c("sol", "BL", "NIST", "CP", "PO"),
  cycle.size = 8,
  rand.method = "none",
  plate.x = as.character(1:9),
  plate.y = c("A", "B", "C", "D", "E", "F"),
  plate.nr = 6,
  one.vial = c("sol"),
  plate.separate = NULL,
  nr.subj.per.block = 0,
  adjust.thres = 0.5,
  sample.position.fixed = FALSE,
  number.reverse = FALSE,
  number.side = FALSE
)
}
\arguments{
\item{dat}{A data set object}

\item{id}{Scope (e.g., country codes or individual IDs)}

\item{time}{Time (e.g., time periods are given by years, months, ...)}
}
\value{
A data frame object that contains a summary of a sample that
    can later be converted to a TeX output using \code{overview_print}
}
\description{
Provides an overview table for the time and scope conditions of
    a data set
}
