% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_lipid_from_lipidmaps.R
\name{get_lipid_from_lipidmaps}
\alias{get_lipid_from_lipidmaps}
\title{get_lipid_from_lipidmaps}
\usage{
get_lipid_from_lipidmaps(lipids = NULL, type = "abbrev", keep.all = FALSE)
}
\arguments{
\item{lipids}{= NULL,}

\item{type}{= "abbrev" # "abbrev" "abbrev_chains"}

\item{keep.all}{= FALSE # keep all records for isomers}
}
\value{
A data frame object that contains a summary of a sample that
    can later be converted to a TeX output using \code{overview_print}
}
\description{
Provides an overview table for the time and scope conditions of
    a data set
}
