% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_PCA.R
\name{ggbiplot}
\alias{ggbiplot}
\title{ggbiplot}
\usage{
ggbiplot(
  pcobj,
  plotly.text = rownames(pcobj$x),
  plot.type = NULL,
  choices = 1:2,
  scale = 1,
  pc.biplot = TRUE,
  obs.scale = 1 - scale,
  var.scale = scale,
  groups = NULL,
  ellipse = FALSE,
  ellipse.prob = 0.68,
  labels = NULL,
  labels.size = 3,
  alpha = 1,
  var.axes = TRUE,
  circle = FALSE,
  circle.prob = 0.69,
  varname.size = 3,
  varname.adjust = 1.5,
  varname.abbrev = FALSE,
  text.size = text.size,
  ...
)
}
\arguments{
\item{dat}{A data set object}

\item{id}{Scope (e.g., country codes or individual IDs)}

\item{time}{Time (e.g., time periods are given by years, months, ...)}
}
\value{
A data frame object that contains a summary of a sample that
    can later be converted to a TeX output using \code{overview_print}
}
\description{
Provides an overview table for the time and scope conditions of
    a data set
}
