% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_various_methods.R
\name{impute_various_methods}
\alias{impute_various_methods}
\title{impute_various_methods}
\usage{
impute_various_methods(x, method = NULL, verbose = FALSE)
}
\arguments{
\item{x}{Input data frame}

\item{method}{= NULL,}

\item{verbose}{= FALSE}
}
\value{
A data frame object that contains a summary of a sample that
    can later be converted to a TeX output using \code{overview_print}
}
\description{
Provides an overview table for the time and scope conditions of
    a data set
}
