% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_filename_to_meta.R
\name{map_filename_to_meta}
\alias{map_filename_to_meta}
\title{Map file names to meta}
\usage{
map_filename_to_meta(raw_files)
}
\arguments{
\item{raw_files}{}
}
\description{
Map file names to meta
}
