% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_metabolite_info.R
\name{map_metabolite_info}
\alias{map_metabolite_info}
\title{map_metabolite_info}
\usage{
map_metabolite_info(
  input_file = NULL,
  name_col = 6,
  sep = ";",
  db_file = NULL,
  db_sep = ",",
  path_output = NULL
)
}
\arguments{
\item{input_file}{A feature table in dataframe, xlsx or csv format. contains at least one
columns: metabolite name}

\item{name_col}{Position of the column containing metabolite names}

\item{sep}{Separator in input_file if it is csv format}

\item{db_file}{Database file to use for mapping metabolite info in csv format}

\item{db_sep}{Separator in db_file in csv format}

\item{path_output}{The path to store output .xlsx file}
}
\value{
A feature table with metabolite information
}
\description{
map_metabolite_info
}
