% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_standard_name_to_meta.R
\name{map_standard_name_to_meta}
\alias{map_standard_name_to_meta}
\title{Map file names to meta}
\usage{
map_standard_name_to_meta(
  raw_files,
  qc_types = c("BL", "NIST", "PO", "sol", "CP", "IQ", "BPL", "MMix")
)
}
\arguments{
\item{raw_files}{}
}
\description{
Map file names to meta
}
