% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_library.R
\name{match_library}
\alias{match_library}
\title{match_library}
\usage{
match_library(
  feature_info = NULL,
  rt_tolerance = 0.1,
  mz_tolerance = 0.01,
  library_file = NULL,
  mode = NULL,
  output.file = NULL
)
}
\arguments{
\item{feature_info}{A dataframe or xlsx of features with column mz and rt}

\item{rt_tolerance}{Retention time tolerance for matching, default is 0.1}

\item{mz_tolerance}{m/z tolerance for matching, default is 0.01}

\item{library_file}{A library file in dataframe or xlsx format. mz, rt and 
compound names are in 1st, 2nd and 2rd column.}

\item{mode}{Polarity, 'pos' or 'neg'}

\item{output.file}{Path and name for the output file}
}
\value{
A data frame contains featured annotated with library
}
\description{
Match features with library based on m/z and RT
}
