% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_library_id.R
\name{match_library_id}
\alias{match_library_id}
\title{match_library_id}
\usage{
match_library_id(peaklist = NULL, lib = NULL)
}
\arguments{
\item{peaklist}{A xlsx of features with column "Identity}

\item{lib}{A library file in dataframe format. It must contain a column named 
"Library.name"}
}
\value{
A data frame of features matched with library
}
\description{
Match features with library based on identity
}
