% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_impute.R
\name{missing_impute}
\alias{missing_impute}
\title{missing_impute}
\usage{
missing_impute(
  datalist = NULL,
  missing.impute.method.sample = NULL,
  missing.impute.method.is = NULL
)
}
\arguments{
\item{datalist}{= NULL,}

\item{missing.impute.method.sample}{= NULL,}

\item{missing.impute.method.is}{= NULL}
}
\value{
A data frame object that contains a summary of a sample that
    can later be converted to a TeX output using \code{overview_print}
}
\description{
Provides an overview table for the time and scope conditions of
    a data set
}
