% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{normalization}
\alias{normalization}
\title{normalization}
\usage{
normalization(
  datalist = NULL,
  impute.method.sample = NULL,
  impute.method.is = NULL,
  po.sample.to.use = NULL,
  norm.method = NULL,
  sample.type.keep = "Sample"
)
}
\arguments{
\item{datalist}{}

\item{impute.method.sample}{}

\item{impute.method.is}{}

\item{po.sample.to.use}{}

\item{norm.method}{}

\item{sample.type.keep}{}
}
\description{
normalization
}
