% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_with_best_internal_standard.R
\name{normalize_with_best_internal_standard}
\alias{normalize_with_best_internal_standard}
\title{normalize_with_best_internal_standard}
\usage{
normalize_with_best_internal_standard(
  x,
  istds,
  batch,
  batch.wise = TRUE,
  cor.method = "pearson",
  type = NULL,
  use.type = NULL,
  verbose = FALSE
)
}
\arguments{
\item{dat}{A data set object}

\item{id}{Scope (e.g., country codes or individual IDs)}

\item{time}{Time (e.g., time periods are given by years, months, ...)}
}
\value{
A data frame object that contains a summary of a sample that
    can later be converted to a TeX output using \code{overview_print}
}
\description{
Provides an overview table for the time and scope conditions of
    a data set
}
