% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_forest.R
\name{plot_forest}
\alias{plot_forest}
\title{plot_forest}
\usage{
plot_forest(
  data = NULL,
  contrast.to.use = NULL,
  contrast.to.sort = NULL,
  coef.to.use = NULL,
  coef.to.sort = NULL,
  alpha = 0.8,
  yintercept = 0,
  Significance = NULL,
  point.shape = 19,
  top.nr = NULL,
  path.result = NULL,
  fig.width = 8,
  fig.height = 6,
  map.contrast = NULL,
  color.manual = c("#4682B4", "#B4464B", "#B4AF46")
)
}
\arguments{
\item{data}{= NULL,}

\item{contrast.to.use}{= NULL,}

\item{contrast.to.sort}{= NULL,}

\item{coef.to.use}{= NULL,}

\item{coef.to.sort}{= NULL,}

\item{alpha}{= 0.8,}

\item{yintercept}{= 0,}

\item{Significance}{= NULL,}

\item{point.shape}{= 19,}

\item{top.nr}{= NULL,}

\item{path.result}{= NULL,}

\item{fig.width}{= 8,}

\item{fig.height}{= 6,}

\item{map.contrast}{= NULL,}

\item{color.manual}{= c("#4682B4", "#B4464B", "#B4AF46")}
}
\value{
A plot
}
\description{
plot forest plot
}
