% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_forest_classes.R
\name{plot_forest_classes}
\alias{plot_forest_classes}
\title{plot_forest_classes}
\usage{
plot_forest_classes(
  data = NULL,
  y.log = FALSE,
  cut.off.1 = 0.05,
  cut.off.2 = 0.001,
  cut.off.label = 1e-07,
  alpha = 0.8,
  format.adjust = TRUE,
  max.overlaps = 10
)
}
\arguments{
\item{data}{= NULL,}

\item{y.log}{= FALSE, # "log" "log2"}

\item{cut.off.1}{= 0.05,}

\item{cut.off.2}{= 0.001,}

\item{cut.off.label}{= 1e-7,}

\item{alpha}{= 0.8,}

\item{format.adjust}{= TRUE,}

\item{max.overlaps}{= 10}
}
\value{
A plot
}
\description{
plot lipidomics forest plot
}
