% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_istd_rsd.R
\name{plot_istd_rsd}
\alias{plot_istd_rsd}
\title{plot_istd_rsd}
\usage{
plot_istd_rsd(datalist = NULL)
}
\arguments{
\item{dat}{A data set object}

\item{id}{Scope (e.g., country codes or individual IDs)}

\item{time}{Time (e.g., time periods are given by years, months, ...)}
}
\value{
A data frame object that contains a summary of a sample that
    can later be converted to a TeX output using \code{overview_print}
}
\description{
Provides an overview table for the time and scope conditions of
    a data set
}
