% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_PCA.R
\name{plot_PCA}
\alias{plot_PCA}
\title{plot_PCA}
\usage{
plot_PCA(
  x,
  group = NULL,
  group.label = "Group",
  plotly.text = rownames(x),
  alpha = 0.5,
  print = FALSE,
  plot.type = "score",
  ratio = 1,
  color.manual = NULL,
  color.scheme = NULL,
  xlim = NULL,
  ylim = NULL,
  pcs = 1:2,
  text.size = 3,
  circle = TRUE,
  ellipse = TRUE,
  ...
)
}
\arguments{
\item{dat}{A data set object}

\item{id}{Scope (e.g., country codes or individual IDs)}

\item{time}{Time (e.g., time periods are given by years, months, ...)}
}
\value{
A data frame object that contains a summary of a sample that
    can later be converted to a TeX output using
}
\description{
Provides an overview table for the time and scope conditions of
    a data set
}
\examples{
data(toydata)
output_table <- overview_tab(dat = toydata, id = ccode, time = year)
}
