% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_PCA_Q_T2.R
\name{plot_PCA_Q_T2}
\alias{plot_PCA_Q_T2}
\title{plot_PCA_Q_T2}
\usage{
plot_PCA_Q_T2(
  x,
  id = NULL,
  highlight = NULL,
  alpha = 0.5,
  plotly.text = rownames(x)
)
}
\arguments{
\item{id}{Scope (e.g., country codes or individual IDs)}

\item{dat}{A data set object}

\item{time}{Time (e.g., time periods are given by years, months, ...)}
}
\value{
A data frame object that contains a summary of a sample that
    can later be converted to a TeX output using \code{overview_print}
}
\description{
Provides an overview table for the time and scope conditions of
    a data set
}
\examples{
data(toydata)
output_table <- overview_tab(dat = toydata, id = ccode, time = year)
}
