% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_concentrations.R
\name{predict_x}
\alias{predict_x}
\title{predict_x}
\usage{
predict_x(y = NULL, mod = NULL, x_interval = NULL)
}
\arguments{
\item{dat}{A data set object}

\item{id}{Scope (e.g., country codes or individual IDs)}

\item{time}{Time (e.g., time periods are given by years, months, ...)}
}
\value{
A data frame object
}
\description{
Provides an overview table for the time and scope conditions of
    a data set
}
