% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_mzmine_xml.R
\name{prepare_mzmine_xml}
\alias{prepare_mzmine_xml}
\title{prepare_mzmine_xml}
\usage{
prepare_mzmine_xml(
  type = NULL,
  path_mzML = NULL,
  path_output = NULL,
  mode = NULL,
  path_lib = NULL,
  mzmine_mz_tol = NULL,
  mzmine_mz_tol_bigger = NULL,
  mzmine_rt_tol = NULL,
  mzmine_rt_tol_bigger = NULL,
  mzmine_min_peaks_in_row = NULL,
  mzmine_min_peaks_in_row_percent = NULL
)
}
\arguments{
\item{type}{= NULL,}

\item{path_mzML}{= NULL,}

\item{path_output}{= NULL,}

\item{mode}{= NULL,}

\item{path_lib}{= NULL,}

\item{mzmine_mz_tol}{= NULL,}

\item{mzmine_mz_tol_bigger}{= NULL,}

\item{mzmine_rt_tol}{= NULL,}

\item{mzmine_rt_tol_bigger}{= NULL,}

\item{mzmine_min_peaks_in_row}{= NULL,}

\item{mzmine_min_peaks_in_row_percent}{= NULL}
}
\value{
A list of summaries for all types of samples
}
\description{
Prepare MZmine xml file
}
