% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_peak_group.R
\name{process_peak_group}
\alias{process_peak_group}
\title{process_peak_group}
\usage{
process_peak_group(
  files = NULL,
  para = NULL,
  sample.info = NULL,
  mode = NULL,
  path.result = NULL
)
}
\arguments{
\item{files}{= NULL,}

\item{para}{= NULL,}

\item{sample.info}{= NULL,}

\item{mode}{= NULL,}

\item{path.result}{= NULL}
}
\value{
A data frame object that contains a summary of a sample that
    can later be converted to a TeX output using \code{overview_print}
}
\description{
Provides an overview table for the time and scope conditions of
    a data set
}
