% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_rmarkdown.R
\name{render_rmarkdown}
\alias{render_rmarkdown}
\title{render_rmarkdown}
\usage{
render_rmarkdown(file = NULL, params = NULL, output = NULL)
}
\arguments{
\item{file}{= NULL,}

\item{params}{= NULL,}

\item{output}{= NULL}
}
\value{
A data frame object that contains a summary of a sample that
    can later be converted to a TeX output using \code{overview_print}
}
\description{
Provides an overview table for the time and scope conditions of
    a data set
}
