% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rt_mapping.R
\name{rt_mapping}
\alias{rt_mapping}
\title{rt_mapping}
\usage{
rt_mapping(ref.file, cor_threshold = 0.5, mz_tolerance = 0.01)
}
\arguments{
\item{dat}{A data set object}

\item{id}{Scope (e.g., country codes or individual IDs)}

\item{time}{Time (e.g., time periods are given by years, months, ...)}
}
\value{
A data frame object that contains a summary of a sample that
    can later be converted to a TeX output using \code{overview_print}
}
\description{
Provides an overview table for the time and scope conditions of
    a data set
}
