% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_import_peaktable.R
\name{run_import_peaktable}
\alias{run_import_peaktable}
\title{run_import_peaktable}
\usage{
run_import_peaktable(
  peaktable_pos = NULL,
  peaktable_neg = NULL,
  rt_col_nr = NULL,
  mz_col_nr = NULL,
  identity_col_nr = NULL,
  sample_col_nr = NULL,
  meta_file_pos = NULL,
  meta_file_neg = NULL,
  find_is = FALSE,
  add_lipid_info = FALSE,
  keep.lipid.orig = NULL,
  standerdized.name = FALSE
)
}
\arguments{
\item{peaktable_pos}{}

\item{peaktable_neg}{}

\item{rt_col_nr}{}

\item{mz_col_nr}{}

\item{identity_col_nr}{}

\item{sample_col_nr}{}

\item{meta_file_pos}{}

\item{meta_file_neg}{}

\item{find_is}{}

\item{add_lipid_info}{}

\item{keep.lipid.orig}{}
}
\description{
run_import_peaktable
}
