% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_precursor.R
\name{select_precursor}
\alias{select_precursor}
\title{select_precursor}
\usage{
select_precursor(
  XCMSnExp = NULL,
  method = NULL,
  mode = NULL,
  remove_frag_addu = FALSE,
  path.result = NULL,
  BPPARAM = NULL
)
}
\arguments{
\item{XCMSnExp}{A XCMSnExp object from XCMS}

\item{method}{The method for annotate and group features, 'camera' or 'pmd'}

\item{mode}{Polarity, 'pos' or 'neg'}

\item{remove_frag_addu}{Indicator for removing the fragments and adducts. TRUE or FALSE}

\item{path.result}{Path to output folder}
}
\value{
A data frame object that contains the precusors extracted from the 
   XCMSnExp object
}
\description{
Select precursors from a XCMSnExp object
}
