% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XCMSnExp_mzrt.R
\name{XCMSnExp_mzrt}
\alias{XCMSnExp_mzrt}
\title{XCMSnExp_mzrt}
\usage{
XCMSnExp_mzrt(
  XCMSnExp,
  method = "medret",
  value = "into",
  mzdigit = 4,
  rtdigit = 1
)
}
\arguments{
\item{XCMSnExp}{XCMS object?}

\item{method}{= "medret",}

\item{value}{= "into",}

\item{mzdigit}{= 4,}

\item{rtdigit}{= 1}
}
\value{
A data frame object that contains a summary of a sample that
    can later be converted to a TeX output using \code{overview_print}
}
\description{
Provides an overview table for the time and scope conditions of
    a data set
}
