% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_with_library.R
\name{annotate_with_library}
\alias{annotate_with_library}
\title{annotate_with_library}
\usage{
annotate_with_library(
  precursor = NULL,
  method = NULL,
  mode = NULL,
  rt_tolerance = NULL,
  mz_tolerance = NULL,
  path.lib = NULL,
  path.result = NULL
)
}
\arguments{
\item{precursor}{= NULL,}

\item{method}{= NULL,}

\item{mode}{= NULL,}

\item{rt_tolerance}{= NULL,}

\item{mz_tolerance}{= NULL,}

\item{path.lib}{= NULL,}

\item{path.result}{= NULL}
}
\value{
A data frame object that contains a summary of a sample that
    can later be converted to a TeX output using \code{overview_print}
}
\description{
Provides an overview table for the time and scope conditions of
    a data set
}
