% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_add_rsd.R
\name{calculate_add_rsd}
\alias{calculate_add_rsd}
\title{calculate_add_rsd}
\usage{
calculate_add_rsd(data = NULL, type = NULL)
}
\arguments{
\item{data}{A dataframe of peaktable, sample x feature}

\item{type}{A vector indicating sample type}
}
\value{
A dataframe object that contains a summary of features (mean and rsd)
}
\description{
Calculate mean and rsd of features in a format to add to the
             feature info later
}
