% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_lm.R
\name{compute_lm}
\alias{compute_lm}
\title{compute_lm}
\usage{
compute_lm(
  data = NULL,
  formula = NULL,
  interaction = NULL,
  covariates = NULL,
  random.effect = NULL,
  dv = NULL,
  path.result = NULL,
  map.names = NULL,
  scaling = TRUE
)
}
\arguments{
\item{data}{= NULL,}

\item{formula}{= NULL,}

\item{interaction}{= NULL,}

\item{covariates}{= NULL,}

\item{random.effect}{= NULL,}

\item{dv}{= NULL,}

\item{path.result}{= NULL,}

\item{map.names}{= NULL}
}
\value{
A dataframe
}
\description{
Provides an overview table for the time and scope conditions of
    a data set
}
