% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_lmm.R
\name{compute_lmm}
\alias{compute_lmm}
\title{compute_lmm}
\usage{
compute_lmm(
  data = NULL,
  formula = NULL,
  interaction = NULL,
  random.effect = NULL,
  dv = NULL,
  output.file = NULL,
  map.names = NULL,
  scaling = TRUE
)
}
\arguments{
\item{data}{= NULL,}

\item{formula}{= NULL,}

\item{interaction}{= NULL,}

\item{random.effect}{= NULL,}

\item{dv}{= NULL,}

\item{map.names}{= NULL}

\item{covariates}{= NULL,}

\item{path.result}{= NULL,}
}
\value{
A dataframe
}
\description{
Provides an overview table for the time and scope conditions of
    a data set
}
