% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.qc.R
\name{create.qc}
\alias{create.qc}
\title{create.qc}
\usage{
create.qc(sampled_data = NULL, figure_height = NULL)
}
\arguments{
\item{sampled_data}{A dataframe containing IS in samples}

\item{figure_height}{Figure height for plot}
}
\value{
A list qc figures
}
\description{
Create qc plots for internal standards
}
