% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_qc.R
\name{create_qc}
\alias{create_qc}
\title{create_qc}
\usage{
create_qc(sampled_data = NULL, figure_height = 300, figure_width = 600)
}
\arguments{
\item{sampled_data}{A dataframe containing IS in samples}

\item{figure_height}{Figure height for plot}

\item{figure_width}{Figure width for plot}
}
\value{
A list qc figures
}
\description{
Create qc plots for internal standards
}
