% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.IS.R
\name{extract.IS}
\alias{extract.IS}
\title{extract.IS}
\usage{
extract.IS(
  sample.info = NULL,
  reference.type = NULL,
  lib.istd = NULL,
  path.peaktable.IS = NULL,
  path.result = NULL
)
}
\arguments{
\item{sample.info}{= NULL,}

\item{reference.type}{= NULL,}

\item{lib.istd}{= NULL,}

\item{path.peaktable.IS}{= NULL,}

\item{path.result}{= NULL}
}
\value{
A list of summaries for all types of samples
}
\description{
Calculate summary for subsets of dataset
}
