% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.is.R
\name{extract.is}
\alias{extract.is}
\title{extract.is}
\usage{
extract.is(
  sample.info = NULL,
  reference.type = NULL,
  lib.istd = NULL,
  path.peaktable.IS = NULL,
  path.result = NULL
)
}
\arguments{
\item{sample.info}{A dataframe containing sample info}

\item{reference.type}{mz and RT to compare to, choose from c("reference", "median")}

\item{lib.istd}{A dataframe of library for internal standards}

\item{path.peaktable.IS}{Path to peaktable for internal standards in samples}

\item{path.result}{Path to output files}
}
\value{
A list of summaries for all types of samples
}
\description{
Extract internal standard measurements from MZmine3 output
}
