% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_feature_info.R
\name{extract_feature_info}
\alias{extract_feature_info}
\title{extract_feature_info}
\usage{
extract_feature_info(
  files = NULL,
  lip_names = NULL,
  met_names = NULL,
  mz_col_nr = 1,
  rt_col_nr = 2,
  identity_col_nr = 3,
  library_file = NULL
)
}
\arguments{
\item{files}{A dataframe of peaktable}

\item{mz_col_nr}{Column position of mz}

\item{rt_col_nr}{Column position of rt}

\item{identity_col_nr}{Column position of identity}

\item{library_file}{A dataframe of library}
}
\value{
A data frame object that contains feature info
}
\description{
Provides an overview table for the time and scope conditions of
    a data set
}
