% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_sample_info.R
\name{extract_sample_info}
\alias{extract_sample_info}
\title{extract_sample_info}
\usage{
extract_sample_info(
  path = NULL,
  files = NULL,
  sample.pattern = NULL,
  standard.name = TRUE,
  path.meta = NULL,
  meta.match.col = "Sample.name",
  group.var = NULL,
  qc.sample.type = c("sol", "BL", "CP", "BP", "NIST", "CAL"),
  calibration.sample.type = c("CA00", "CA01", "CA02", "CA03", "CA04", "CA05", "CA06",
    "CA07", "CA08", "CA09"),
  export.rds = NULL,
  export.xlsx = NULL
)
}
\arguments{
\item{path}{Path to converted data files}

\item{files}{converted data file names}

\item{sample.pattern}{Sample file format, e.g. ".mzML"}

\item{standard.name}{If standardized sample names are used, TRUE/FALSE}

\item{path.meta}{Metadata file, which should have a key column to match with
sample names}

\item{meta.match.col}{Column used as key column for matching}

\item{group.var}{The name of the variable indicating groups of samples}

\item{qc.sample.type}{QC sample types}

\item{calibration.sample.type}{Calibration sample types}

\item{export.rds}{rds name to export}

\item{export.xlsx}{xlsx name to export}
}
\value{
A data frame object that contains sample names and related information
}
\description{
Extract sample information from the standardized sample names or 
   metadata and generate sample list
}
