% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_peaks.R
\name{filter_peaks}
\alias{filter_peaks}
\title{filter_peaks}
\usage{
filter_peaks(
  XCMSnExp = NULL,
  peaktable = NULL,
  sample.type = NULL,
  bl.thres = NULL,
  rsd.po.thres = NULL,
  rt.range = NULL,
  mean.thres = NULL
)
}
\arguments{
\item{XCMSnExp}{= NULL,}

\item{peaktable}{= NULL,}

\item{sample.type}{= NULL,}

\item{bl.thres}{= NULL,}

\item{rsd.po.thres}{= NULL,}

\item{rt.range}{= NULL,}

\item{mean.thres}{= NULL}
}
\value{
A data frame object that contains a summary of a sample that
    can later be converted to a TeX output using \code{overview_print}
}
\description{
Provides an overview table for the time and scope conditions of
    a data set
}
\examples{
data(toydata)
output_table <- overview_tab(dat = toydata, id = ccode, time = year)
}
