% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_mzmine3_xml.R
\name{generate_mzmine3_xml}
\alias{generate_mzmine3_xml}
\title{generate_mzmine3_xml}
\usage{
generate_mzmine3_xml(template = NULL, params = NULL, output_file = NULL)
}
\arguments{
\item{template}{= NULL,}

\item{params}{= NULL,}

\item{output_file}{= NULL}
}
\value{
A data frame object
}
\description{
Provides an overview table for the time and scope conditions of
    a data set
}
