% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_peaktable.R
\name{import_peaktable}
\alias{import_peaktable}
\title{import_peaktable}
\usage{
import_peaktable(
  peaktable = NULL,
  meta = NULL,
  rt_col_nr = NULL,
  mz_col_nr = NULL,
  identity_col_nr = NULL,
  sample_col_nr = NULL,
  find_is = FALSE,
  add_lipid_info = FALSE,
  keep.lipid.orig = NULL,
  standerdized.name = FALSE
)
}
\arguments{
\item{dat}{A data set object}

\item{id}{Scope (e.g., country codes or individual IDs)}

\item{time}{Time (e.g., time periods are given by years, months, ...)}
}
\value{
A data frame object that contains a summary of a sample that
    can later be converted to a TeX output using \code{overview_print}
}
\description{
Provides an overview table for the time and scope conditions of
    a data set
}
\examples{
data(toydata)
output_table <- overview_tab(dat = toydata, id = ccode, time = year)
}
