% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_library.R
\name{match_library}
\alias{match_library}
\title{match_library}
\usage{
match_library(
  feature_info = NULL,
  rt_tolerance = 0.1,
  mz_tolerance = 0.01,
  library_file = NULL,
  mode = NULL,
  output.file = NULL
)
}
\arguments{
\item{dat}{A data set object}

\item{id}{Scope (e.g., country codes or individual IDs)}

\item{time}{Time (e.g., time periods are given by years, months, ...)}
}
\value{
A data frame object that contains a summary of a sample that
    can later be converted to a TeX output using \code{overview_print}
}
\description{
Provides an overview table for the time and scope conditions of
    a data set
}
