% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_various_methods.R
\name{normalize_various_methods}
\alias{normalize_various_methods}
\title{normalize_various_methods}
\usage{
normalize_various_methods(
  x,
  istds = NULL,
  method = NULL,
  type = NULL,
  use.type = NULL,
  reference = NULL,
  use.reference = NULL,
  group = NULL,
  verbose = FALSE,
  batch = NULL,
  batch.wise = FALSE,
  sample.rate = 0.33,
  data.rsd.orig = NULL,
  export.path = NULL,
  prefix = "",
  feature.info = NULL
)
}
\arguments{
\item{dat}{A data set object}

\item{id}{Scope (e.g., country codes or individual IDs)}

\item{time}{Time (e.g., time periods are given by years, months, ...)}
}
\value{
A data frame object that contains a summary of a sample that
    can later be converted to a TeX output using \code{overview_print}
}
\description{
Provides an overview table for the time and scope conditions of
    a data set
}
\examples{
data(toydata)
output_table <- overview_tab(dat = toydata, id = ccode, time = year)
}
