% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_para.R
\name{optim_para}
\alias{optim_para}
\title{optim_para}
\usage{
optim_para(files = NULL, mode = NULL, path.result = NULL, BPPARAM = NULL)
}
\arguments{
\item{files}{= NULL,}

\item{mode}{= NULL,}

\item{path.result}{= NULL}

\item{BPPARAM}{= NULL}
}
\value{
A data frame object that contains a summary of a sample that
    can later be converted to a TeX output using \code{overview_print}
}
\description{
Provides an overview table for the time and scope conditions of
    a data set
}
