% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params_setup.R
\name{params_setup}
\alias{params_setup}
\title{params_setup}
\usage{
params_setup(
  path.mzml = NULL,
  standard.name = TRUE,
  path.meta = NULL,
  meta.match.col = "Sample",
  qc.sample.type = c("sol", "BL", "CP", "IQ", "NIST", "CAL", "Met", "M-Mix", "SPO", "PO",
    "PO50", "PO12.5", "PO25", "PO100", "BPO"),
  calibration.sample.type = c("CA00", "CA01", "CA02", "CA03", "CA04", "CA05", "CA06",
    "CA07", "CA08", "CA09"),
  rt_tolerance = 0.3,
  mz_tolerance = 0.01,
  type.use.to.optimize = "PO",
  po.sample.to.use = "PO",
  path.mzmine = NULL,
  path.msdial = NULL,
  BPPARAM_set = NULL,
  method = NULL,
  path.result = NULL,
  sample.pattern = ".mzML",
  path.lib = NULL,
  path.lib.is = NULL,
  path.msp = NULL,
  path.msp.pos = NULL,
  path.msp.neg = NULL
)
}
\arguments{
\item{path.mzml}{Path to mzML files}

\item{standard.name}{If standardized sample names are used, TRUE/FALSE}

\item{path.meta}{Path to metadata file for samples}

\item{meta.match.col}{Column name to match sample info and metadata, default is "Sample"

Specification sample types used for QC and calibration}

\item{qc.sample.type}{QC sample types, default is c( "sol", "BL", "CP", "IQ",
"NIST", "CAL", "Met", "M-Mix", "SPO", "PO", "PO50", "PO12.5", "PO25", "PO100", "BPO" )}

\item{calibration.sample.type}{Calibration sample types, default is c( "CA00",
   "CA01", "CA02","CA03", "CA04", "CA05", "CA06", "CA07", "CA08", "CA09" )

Parameters for preprocessing}

\item{rt_tolerance}{Retention time tolerance for matching library}

\item{mz_tolerance}{m/z tolerance for matching library}

\item{type.use.to.optimize}{Sample type used for optimizing XCMS parameters, default is "PO"}

\item{po.sample.to.use}{Sample type used for evaluating normalization, default is "PO"

Path to preprocessing software}

\item{path.mzmine}{Path to MZmine3 software, default is}

\item{path.msdial}{Path to MSDIAL ConsoleApp}

\item{BPPARAM_set}{Parallel processing setup

Other parameters: no need to change for standard workflow}

\item{method}{Data acqusition method, choose from c("LIP", "RP", "HILIC")}

\item{path.result}{Path to output folder}

\item{sample.pattern}{Sample pattern of data files, default is ".mzML"

Path to library}

\item{path.lib}{Path to in house library in rds/xlsx format. In xlsx format, 
sheets are labelled as POS and NEG}

\item{path.lib.is}{Path to in house library for internal standards in rds/xlsx format.
In xlsx format, sheets are labelled as POS and NEG}

\item{path.msp}{Path to public MS2 database file in rds

If the rds file for public MS2 database is not available, msp format can be provided}

\item{path.msp.pos}{Path to public MS2 positive mode database file in msp format}

\item{path.msp.neg}{Path to public MS2 negative mode database file in msp format}
}
\value{
A list of parameters used for data processing
}
\description{
Set up parameters for data processing

Path to data files
}
