% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params_setup.R, R/params_setup_mzmine2.R
\name{params_setup}
\alias{params_setup}
\title{params_setup}
\usage{
params_setup(
  params = NULL,
  path.mzmine = NULL,
  path.msdial = NULL,
  BPPARAM_set = NULL,
  qc.sample.type = c("sol", "BL", "CP", "IQ", "NIST", "CAL", "Met", "M-Mix", "SPO", "PO",
    "PO50", "PO12.5", "PO25", "PO100", "BPO"),
  calibration.sample.type = c("CA00", "CA01", "CA02", "CA03", "CA04", "CA05", "CA06",
    "CA07", "CA08", "CA09"),
  path.mzml = NULL,
  standard.name = TRUE,
  target = NULL,
  path.meta.pos = NULL,
  path.meta.neg = NULL,
  meta.match.col = "Sample",
  rt_tolerance = 0.3,
  mz_tolerance = 0.01,
  type.use.to.optimize = "PO",
  po.sample.to.use = "PO",
  method = NULL,
  path.result = NULL,
  sample.pattern = ".mzML",
  path.mzml.pos = NULL,
  path.mzml.neg = NULL,
  path.lib = NULL,
  path.lib.is = NULL,
  path.msp = NULL
)

params_setup(
  params = NULL,
  path.mzmine = NULL,
  path.msdial = NULL,
  BPPARAM_set = NULL,
  qc.sample.type = c("sol", "BL", "CP", "IQ", "NIST", "CAL", "Met", "M-Mix", "SPO", "PO",
    "PO50", "PO12.5", "PO25", "PO100", "BPO"),
  calibration.sample.type = c("CA00", "CA01", "CA02", "CA03", "CA04", "CA05", "CA06",
    "CA07", "CA08", "CA09"),
  path.mzml = NULL,
  standard.name = TRUE,
  target = NULL,
  path.meta.pos = NULL,
  path.meta.neg = NULL,
  meta.match.col = "Sample",
  rt_tolerance = 0.3,
  mz_tolerance = 0.01,
  type.use.to.optimize = "PO",
  po.sample.to.use = "PO",
  method = NULL,
  path.result = NULL,
  sample.pattern = ".mzML",
  path.mzml.pos = NULL,
  path.mzml.neg = NULL,
  path.lib = NULL,
  path.lib.is = NULL,
  path.msp = NULL
)
}
\arguments{
\item{params}{A list of parameters, set as NULL if nothing has been set up yet}

\item{path.mzmine}{Path of MZmine}

\item{BPPARAM_set}{Parallel preprocessing setup in XCMS}

\item{qc.sample.type}{QC sample types, default is c( "sol", "BL", "CP", "IQ",
"NIST", "CAL", "Met", "M-Mix", "SPO", "PO", "PO50", "PO12.5", "PO25", "PO100", "BPO" )}

\item{calibration.sample.type}{Calibration sample types, default is c( "CA00",
"CA01", "CA02","CA03", "CA04", "CA05", "CA06", "CA07", "CA08", "CA09" )}

\item{path.mzml}{Path to mzML files}

\item{standard.name}{If standardized sample names are used, TRUE/FALSE}

\item{path.meta.pos}{Path to metadata file for samples from positive mode}

\item{path.meta.neg}{Path to metadata file for samples from negative mode}

\item{meta.match.col}{Column name to match sample and metadata, "Sample" for
standard.name = TRUE, "Sample.name" for standard.name = FALSE}

\item{rt_tolerance}{Retention time tolerance for matching library}

\item{mz_tolerance}{m/z for matching library}

\item{type.use.to.optimize}{Sample type used for optimizing XCMS parameters, default is "PO"}

\item{po.sample.to.use}{Sample type used for evaluating normalization, default is "PO"}

\item{method}{Choose from c("LIP", "RP", "HILIC")}

\item{path.result}{Path to result folder}

\item{sample.pattern}{Sample pattern of data files, default is ".mzML"}

\item{path.mzml.pos}{Path to mzML positive mode files, put "No_data" if not preprocessing the mode}

\item{path.mzml.neg}{Path to mzML negative mode files, put "No_data" if not preprocessing the mode}

\item{path.lib}{Path to in house library xlsx/RData}

\item{path.lib.is}{Path to in house library for internal standards xlsx/RData}

\item{run.pipeline}{If run whole pipeline or not, TRUE/FALSE}
}
\value{
A list of parameters

A list of parameters
}
\description{
Set up preprocessing parameters

Set up preprocessing parameters
}
