% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_datalist.R
\name{prepare_datalist}
\alias{prepare_datalist}
\title{prepare_datalist}
\usage{
prepare_datalist(
  peaks = NULL,
  sample.info = NULL,
  sample.pattern = "",
  type.to.remove = c("Met", "CA")
)
}
\arguments{
\item{dat}{A data set object}

\item{id}{Scope (e.g., country codes or individual IDs)}

\item{time}{Time (e.g., time periods are given by years, months, ...)}
}
\value{
A data frame object that contains a summary of a sample that
    can later be converted to a TeX output using \code{overview_print}
}
\description{
Provides an overview table for the time and scope conditions of
    a data set
}
