% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_mzmine_xml.R
\name{prepare_mzmine_xml}
\alias{prepare_mzmine_xml}
\title{prepare_mzmine_xml}
\usage{
prepare_mzmine_xml(
  type = NULL,
  path_mzML = NULL,
  path_output = NULL,
  mode = NULL,
  path_lib = NULL,
  mzmine_mz_tol = NULL,
  mzmine_mz_tol_bigger = NULL,
  mzmine_rt_tol = NULL,
  mzmine_rt_tol_bigger = NULL,
  mzmine_min_peaks_in_row = NULL,
  mzmine_min_peaks_in_row_percent = NULL
)
}
\arguments{
\item{type}{The type of preprocessing the xml will do. "is" or "peak" for preprocessing
internal standards or all metabolites.}

\item{path_mzML}{Path to the folder containing mzML files}

\item{path_output}{Path to the output folder}

\item{mode}{Polarity, 'pos' or 'neg'}

\item{path_lib}{Path to the library file}

\item{mzmine_mz_tol}{m/z tolerance for feature detection}

\item{mzmine_mz_tol_bigger}{m/z tolerance for alignment}

\item{mzmine_rt_tol}{Retention time tolerance for feature detection}

\item{mzmine_rt_tol_bigger}{Retention time tolerance for alignment}

\item{mzmine_min_peaks_in_row}{Minimum number of samples containing the peak}

\item{mzmine_min_peaks_in_row_percent}{Minimum percentage of samples containing the peak}
}
\value{
A xml file for MZMine processing
}
\description{
General MZmine xml file
}
