% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_peak_group.R
\name{process_peak_group}
\alias{process_peak_group}
\title{process_peak_group}
\usage{
process_peak_group(
  para = NULL,
  sample.info = NULL,
  mode = NULL,
  path.result = NULL,
  BPPARAM = NULL
)
}
\arguments{
\item{para}{A list of parameters used for preprocessing}

\item{sample.info}{Sample information used for preprocessing}

\item{mode}{Polarity, 'pos' or 'neg'}

\item{path.result}{Path to output folder}

\item{BPPARAM}{Parallel processing setup}
}
\value{
A XCMSnExp object from XCMS
}
\description{
Preprocessing with XCMS
}
