% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_merge_and_map.R
\name{run_merge_amd_map}
\alias{run_merge_amd_map}
\title{run_merge_amd_map}
\usage{
run_merge_amd_map(
  datalist = NULL,
  final.norm = NULL,
  eval.sample.to.use = NULL,
  sample.type.keep = NULL,
  db_file = NULL
)
}
\arguments{
\item{datalist}{}

\item{final.norm}{}

\item{eval.sample.to.use}{}

\item{sample.type.keep}{}
}
\description{
run_merge_amd_map
}
