% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_module.R
\name{run_module}
\alias{run_module}
\title{run_module}
\usage{
run_module(module, params, file = NULL)
}
\arguments{
\item{module}{Module name to run}

\item{params}{List of parameters to use}
}
\value{
Rmarkdown files and related output files
}
\description{
Wrapper function to run modules
}
