% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_normalization.R
\name{run_normalization}
\alias{run_normalization}
\title{run_normalization}
\usage{
run_normalization(
  datalist = NULL,
  impute.method.sample = NULL,
  impute.method.is = NULL,
  po.sample.to.use = NULL,
  norm.method = NULL,
  sample.type.keep = "Sample"
)
}
\arguments{
\item{datalist}{}

\item{impute.method.sample}{}

\item{impute.method.is}{}

\item{po.sample.to.use}{}

\item{norm.method}{}

\item{sample.type.keep}{}
}
\description{
run_normalization
}
