% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_precursor.R
\name{select_precursor}
\alias{select_precursor}
\title{select_precursor}
\usage{
select_precursor(XCMSnExp = NULL, mode = NULL, path.result = NULL)
}
\arguments{
\item{XCMSnExp}{= NULL,}

\item{mode}{= NULL,}

\item{path.result}{= NULL}
}
\value{
A data frame object that contains a summary of a sample that
    can later be converted to a TeX output using \code{overview_print}
}
\description{
Provides an overview table for the time and scope conditions of
    a data set
}
