% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/XCMSnExp_mzrt.R
\name{XCMSnExp_mzrt}
\alias{XCMSnExp_mzrt}
\title{XCMSnExp_mzrt}
\usage{
XCMSnExp_mzrt(
  XCMSnExp,
  method = "medret",
  value = "into",
  mzdigit = 4,
  rtdigit = 1
)
}
\arguments{
\item{XCMSnExp}{A XCMSnExp object}

\item{method}{XCMS featureValues parameter for feature values: specifying the 
method to resolve multipeak mappings within the same sample. "medret", report 
the value for the chromatographic peak closest to the feature's median 
retention time.}

\item{value}{XCMS featureValues parameter: defining which value should be reported
for each feature in each sample. Can be any column of the chromPeaks matrix.
Defaults "into", the integrated peak area is reported}

\item{mzdigit}{Number of digits to keep for m/z}

\item{rtdigit}{Number of digits to keep for rt}
}
\value{
A list of m/z, rt
}
\description{
Extract m/z, rt from a XCMSnExp object
}
